move convert_charge_string method
